Switch to cascaded dependancies design
Move away from all depending on a list of recipes to only depending on
push.

All of the sub recipes depend on each other to produce the same output.
This fixes typing make test without a build before hand.

Note: this is a breaking change if you call into the makefile to just
build without a working test